+++ NEWS.md
@@ -1,6 +1,7 @@
 # nlmixr2rpt 0.2.2 (development version)
 
 * Fixed issue with {} in messages creatining the `safe_text()` function to deal with it.
+* Added support for `ggmatrix` figure objects (from {GGally} package).
 * Fixed issue with processing covariates that set them to NA :)
 
 # nlmixr2rpt 0.2.1